refactor: move more GitHub token related utils to check-token module
