+++ lib/modules/manager/composer/utils.spec.ts
@@ -1,13 +1,10 @@
 import { GlobalConfig } from '../../../config/global';
 import * as hostRules from '../../../util/host-rules';
-import { GitTagsDatasource } from '../../datasource/git-tags';
 import { Lockfile, PackageFile } from './schema';
 import {
   extractConstraints,
-  findGithubToken,
   getComposerArguments,
   requireComposerDependencyInstallation,
-  takePersonalAccessTokenIfPossible,
 } from './utils';
 
 jest.mock('../../datasource');
@@ -299,143 +296,6 @@ describe('modules/manager/composer/utils', () => {
         packages: [{ name: 'symfony/console', version: '5.4.0' }],
       });
       expect(requireComposerDependencyInstallation(lockfile)).toBeFalse();
-    });
-  });
-
-  describe('findGithubToken', () => {
-    it('returns the token string when hostRule match search with a valid personal access token', () => {
-      const TOKEN_STRING = 'ghp_TOKEN';
-      hostRules.add({
-        hostType: GitTagsDatasource.id,
-        matchHost: 'github.com',
-        token: TOKEN_STRING,
-      });
-
-      const foundHostRule = hostRules.find({
-        hostType: GitTagsDatasource.id,
-        url: 'https://github.com',
-      });
-
-      expect(findGithubToken(foundHostRule)).toEqual(TOKEN_STRING);
-    });
-
-    it('returns undefined when no token is defined', () => {
-      hostRules.add({
-        hostType: GitTagsDatasource.id,
-        matchHost: 'github.com',
-      });
-
-      const foundHostRule = hostRules.find({
-        hostType: GitTagsDatasource.id,
-        url: 'https://github.com',
-      });
-      expect(findGithubToken(foundHostRule)).toBeUndefined();
-    });
-
-    it('remove x-access-token token prefix', () => {
-      const TOKEN_STRING_WITH_PREFIX = 'x-access-token:ghp_TOKEN';
-      const TOKEN_STRING = 'ghp_TOKEN';
-      hostRules.add({
-        hostType: GitTagsDatasource.id,
-        matchHost: 'github.com',
-        token: TOKEN_STRING_WITH_PREFIX,
-      });
-
-      const foundHostRule = hostRules.find({
-        hostType: GitTagsDatasource.id,
-        url: 'https://github.com',
-      });
-      expect(findGithubToken(foundHostRule)).toEqual(TOKEN_STRING);
-    });
-  });
-
-  describe('takePersonalAccessTokenIfPossible', () => {
-    it('returns undefined when both token are undefined', () => {
-      const githubToken = undefined;
-      const gitTagsGithubToken = undefined;
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toBeUndefined();
-    });
-
-    it('returns gitTagsToken when both token are PAT', () => {
-      const githubToken = 'ghp_github';
-      const gitTagsGithubToken = 'ghp_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(gitTagsGithubToken);
-    });
-
-    it('returns githubToken is PAT and gitTagsGithubToken is not a PAT', () => {
-      const githubToken = 'ghp_github';
-      const gitTagsGithubToken = 'ghs_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(githubToken);
-    });
-
-    it('returns gitTagsToken when both token are set but not PAT', () => {
-      const githubToken = 'ghs_github';
-      const gitTagsGithubToken = 'ghs_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(gitTagsGithubToken);
-    });
-
-    it('returns gitTagsToken when gitTagsToken not PAT and gitTagsGithubToken is not set', () => {
-      const githubToken = undefined;
-      const gitTagsGithubToken = 'ghs_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(gitTagsGithubToken);
-    });
-
-    it('returns githubToken when githubToken not PAT and gitTagsGithubToken is not set', () => {
-      const githubToken = 'ghs_gitTags';
-      const gitTagsGithubToken = undefined;
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(githubToken);
-    });
-
-    it('take personal assess token over fine grained token', () => {
-      const githubToken = 'ghp_github';
-      const gitTagsGithubToken = 'github_pat_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(githubToken);
-    });
-
-    it('take fine grained token over server to server token', () => {
-      const githubToken = 'github_pat_github';
-      const gitTagsGithubToken = 'ghs_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(githubToken);
-    });
-
-    it('take git-tags fine grained token', () => {
-      const githubToken = undefined;
-      const gitTagsGithubToken = 'github_pat_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(gitTagsGithubToken);
-    });
-
-    it('take git-tags unknown token type when no other token is set', () => {
-      const githubToken = undefined;
-      const gitTagsGithubToken = 'unknownTokenType_gitTags';
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(gitTagsGithubToken);
-    });
-
-    it('take github unknown token type when no other token is set', () => {
-      const githubToken = 'unknownTokenType';
-      const gitTagsGithubToken = undefined;
-      expect(
-        takePersonalAccessTokenIfPossible(githubToken, gitTagsGithubToken)
-      ).toEqual(githubToken);
     });
   });
 });
