+++ lib/modules/manager/composer/utils.ts
@@ -3,11 +3,6 @@ import { quote } from 'shlex';
 import { GlobalConfig } from '../../../config/global';
 import { logger } from '../../../logger';
 import type { HostRuleSearchResult } from '../../../types';
-import {
-  isGithubFineGrainedPersonalAccessToken,
-  isGithubPersonalAccessToken,
-  isGithubServerToServerToken,
-} from '../../../util/check-token';
 import type { ToolConstraint } from '../../../util/exec/types';
 import { api, id as composerVersioningId } from '../../versioning/composer';
 import type { UpdateArtifactsConfig } from '../types';
@@ -115,59 +110,6 @@ export function extractConstraints(
     res.composer = `^${major}.${minor}`;
   }
   return res;
-}
-
-export function findGithubToken(
-  searchResult: HostRuleSearchResult
-): string | undefined {
-  return searchResult?.token?.replace('x-access-token:', '');
-}
-
-export function takePersonalAccessTokenIfPossible(
-  githubToken: string | undefined,
-  gitTagsGithubToken: string | undefined
-): string | undefined {
-  if (gitTagsGithubToken && isGithubPersonalAccessToken(gitTagsGithubToken)) {
-    logger.debug('Using GitHub Personal Access Token (git-tags)');
-    return gitTagsGithubToken;
-  }
-
-  if (githubToken && isGithubPersonalAccessToken(githubToken)) {
-    logger.debug('Using GitHub Personal Access Token');
-    return githubToken;
-  }
-
-  if (
-    gitTagsGithubToken &&
-    isGithubFineGrainedPersonalAccessToken(gitTagsGithubToken)
-  ) {
-    logger.debug('Using GitHub Fine-grained Personal Access Token (git-tags)');
-    return gitTagsGithubToken;
-  }
-
-  if (githubToken && isGithubFineGrainedPersonalAccessToken(githubToken)) {
-    logger.debug('Using GitHub Fine-grained Personal Access Token');
-    return githubToken;
-  }
-
-  if (gitTagsGithubToken) {
-    if (isGithubServerToServerToken(gitTagsGithubToken)) {
-      logger.debug('Using GitHub Server-to-Server token (git-tags)');
-    } else {
-      logger.debug('Using unknown GitHub token type (git-tags)');
-    }
-    return gitTagsGithubToken;
-  }
-
-  if (githubToken) {
-    if (isGithubServerToServerToken(githubToken)) {
-      logger.debug('Using GitHub Server-to-Server token');
-    } else {
-      logger.debug('Using unknown GitHub token type');
-    }
-  }
-
-  return githubToken;
 }
 
 export function isArtifactAuthEnabled(rule: HostRuleSearchResult): boolean {
